fix tests

+++ notebooks/demo.jl
@@ -17,7 +17,7 @@ using Optim
 md"# To define a loss"
 
 # ╔═╡ 937e8aa6-f17a-11ea-1d5d-6f495a29d584
-@i function i_solve!(param::AcousticPropagatorParams, srci::Int64, srcj::Int64,
+@i function i_solve!(param::AcousticPropagatorParams, srci::Int, srcj::Int,
             srcv::Array{T, 1}, c::Array{T, 2},
             tu::Array{T,3}, tφ::Array{T,3}, tψ::Array{T,3}) where T
